+++ src/python/pants/backend/build_files/fix/deprecations/register.py
@@ -1,7 +1,7 @@
 # Copyright 2022 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.backend.project_info.fix.deprecations import renamed_fields_rules, renamed_targets_rules
+from pants.backend.build_files.fix.deprecations import renamed_fields_rules, renamed_targets_rules
 
 
 def rules():
