+++ src/python/pants/bin/BUILD
@@ -13,7 +13,7 @@ python_sources(
 target(
     name="plugins",
     dependencies=[
-        "src/python/pants/backend/project_info/fix/deprecations",
+        "src/python/pants/backend/build_files/fix/deprecations",
         "src/python/pants/backend/build_files/fmt/buildifier",
         "src/python/pants/backend/awslambda/python",
         "src/python/pants/backend/build_files/fmt/black",
